fix(session): add with_agent_id/with_role builder and new_with_identity constructor
Source: issue #499
Type: fix

- Add with_agent_id and with_role builder methods to SessionCheckpoint
- Add new_with_identity constructor to CheckpointManager that auto-fills
  agent_id and role on every checkpoint saved through it
- Split CheckpointManager into checkpoint_manager.rs (349 + 172 lines,
  both under 500-line limit)
- new() now also sets a default identity
  (empty string agent_id, MainAgent role)
